move the files to profile directory
