+++ src/components/profile/UserDetailForm/index.scss
@@ -1,7 +1,7 @@
-@import "../../style/variables";
-@import "../../style/buttons";
-@import "../../style/forms";
-@import "../../style/bootstrapImports";
+@import "../../../style/variables";
+@import "../../../style/buttons";
+@import "../../../style/forms";
+@import "../../../style/bootstrapImports";
 
 .userDetailSection {
   font: inherit;
