+++ src/style/_forms.scss
@@ -1,6 +1,3 @@
-@import "../../style/variables";
-@import "../../style/buttons";
-
 @mixin inputElement(
   $bgColor: $lightColor,
   $hoverColor: $primaryColor,
